# openapi.yaml
openapi: 3.1.0
info:
  title: edu_rppa API
  description: |
    Демонстрационное API для мастер-класса по архитектуре информационных систем.
    Связка Frontend ↔ FastAPI ↔ PostgreSQL.
  version: 1.0.0
servers:
  - url: https://edu-rppa-sk0v4edu.amvera.io
    description: Production (Amvera Cloud)
  - url: http://localhost:8000
    description: Local development
...
paths:
  /api/auth/request-code:
    post:
      tags: [Auth]
      summary: Запросить код авторизации
      description: |
        Отправляет 6-значный код на указанный email.
        Если пользователя с таким email нет — создаёт автоматически.
        В режиме DEV_DELIVERY=true код возвращается в ответе (поле demo_code).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                name:
                  type: string
                  description: Имя пользователя (если не указано — генерируется из email)
                  example: Ivan Petrov
      responses:
        "200":
          description: Код отправлен
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  demo_code:
                    type: string
                    description: Код (только при DEV_DELIVERY=true)
                    example: "482710"
        "400":
          description: Невалидный email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/verify-code:
    post:
      tags: [Auth]
      summary: Подтвердить код и создать сессию
      description: |
        Проверяет код, создаёт сессию и устанавливает cookies:
        - `sid` (HttpOnly) — идентификатор сессии
        - `u` — JSON с email и именем для UI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                code:
                  type: string
                  example: "482710"
      responses:
        "200":
          description: Сессия создана, cookies установлены
          headers:
            Set-Cookie:
              description: "sid и u cookies"
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
        "400":
          description: Неверный или просроченный код
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/admin/query:
    get:
      tags: [Admin]
      summary: Выполнить SQL-запрос (временный)
      description: Защищён секретным ключом. Для отладки и просмотра данных.
      parameters:
        - name: secret
          in: query
          required: true
          schema:
            type: string
        - name: sql
          in: query
          required: true
          schema:
            type: string
          example: "SELECT * FROM users"
      responses:
        "200":
          description: Результат запроса
          content:
            application/json:
              schema:
                type: object
                properties:
                  rows:
                    type: array
                    items:
                      type: object
                  count:
                    type: integer
        "403":
          description: Неверный секрет
        "500":
          description: Ошибка SQL
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "valid email required"